ci(goreleaser): use `concurrency` to only execute 1 job at a time (#3557)
This will make it so that on master, there is only 1 job of
releaser-master executing at a time. There may be another pending one
(if, say, we merge two PRs back to back), and it will be put as
`pending`. It's cancelled if we merge a third PR while the second one is
still `pending`.

For more information:
https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#concurrency