refactor(examples): use std.Address based behaviour (#3580)
## Description

To make contracts more simpler(and low computation for execution), it is
better for contract to manipulate with only `std.Address` rather than
using `users.AddressOrName` and resolving it. ([Discussed with
@moul](https://github.com/gnoswap-labs/gnoswap/pull/465#discussion_r1922735140))

Clients such as gnoeky, gnoweb or adena should perform look up(convert
username to address) and pass address value to contract.