feat(github-bot): avoid triage-pending if there is any tech-staff review  (#3574)
After a first day of experimentation, we determined we don't want the
triage-pending label on PRs that have already received a review from a
tech-staff member.

Some additional changes:

- Split between `ApprovalBy` and `ReviewBy` in reviewer.go, to
distinguish between the requirements that only work with an approval and
those that are more generic.
- Don't request a team review if a member of the team already reviewed
or is requested for review. (The bot was nagging a bit on some PRs
because of this)