fix: remove import cycles (#3304)
Depends on #3323 

- add test in `examples/no_cycles_test.go` to detect import cycles in
stdlibs and examples
- add `matchString` native injection in the `testing` stdlib to avoid a
cycle in `regexp` tests
- remove other import cycles

Go never allows import cycles. Our stdlibs have a lot of import cycles,
and some examples import self which is not allowed in golang either.
Keeping support for import cycles in stdlib and importing self will
require a lot of hacky and weird logic in generic package loading code
so I try to tackle this first.

TODO:

- [x] fix tests
- [x] ~~check cycles with the test stdlibs overlay applied~~ -> be
explicit about the lack of support for modifying imports in testing
stdlibs overlay

---------

Signed-off-by: Norman Meier <norman@samourai.coop>
Signed-off-by: Norman <norman@Normans-MacBook-Pro-2.local>
Signed-off-by: Norman <norman@samourai.coop>
Co-authored-by: Morgan <morgan@morganbaz.com>
Co-authored-by: Norman <norman@Normans-MacBook-Pro-2.local>