feat(boards2): improve public realm functions (#3583)
The PR simplifies the implementation of a couple public boards2 realm
functions and change them to be up to date with the latest permission
handling changes.

These changes also allow a creator of a thread to edit it without
permission, or creators of a reply/comment to delete the reply.

Also creators of a thread are allowed to delete it for now. This
behavior might change in future PRs.