feat(gnoweb): "No render" page/component (#3611)
## Description

Defining a `Render()` function in realms is optional. Currently gnoweb
presents an error if a realm that doesn't have a render func is
requested. This should not be the case.

This PR also adds a VM error, `RenderNotDeclared`, which is to be
returned when `vm/qrender` is called on a realm which does not have a
`Render()` function declared.

I updated the status component to return the following in the
aforementioned case:
<img width="1557" alt="Screenshot 2025-01-25 at 16 30 55"
src="https://github.com/user-attachments/assets/dc82c1d6-d815-4d92-a1a9-7639cbf7bca2"
/>


Also adds another `r/docs` realm mentioning that a render function is
optional in `r/`.