chore(examples): modify pausable (#3628)
## Description

Prevously, the `pausable` object embedded the `ownable` object directly,
and with the intended usage of both packages being the following,
`ownable` functions were being duplicated:

```go
var (
	Ownable = ownable.NewWithAddress(std.Address("xyz"))
	Pausable = pausable.NewFromOwnable(Ownable)
)
```

This PR names the `ownable` inside `pausable` as a private field and
exposes a getter in case someone needs it for some reason. It also
removes the `New()` function, which doesn't make too much sense right
now as the pausable needs to be paired with `ownable` - this might
change later but for now this should be the way.

---------

Signed-off-by: moul <94029+moul@users.noreply.github.com>
Co-authored-by: moul <94029+moul@users.noreply.github.com>