feat(gnovm): support constant evaluation of len and cap on arrays (#3600)
Closes: #3201 

This update introduces the ability to evaluate len and cap for arrays at
preprocess-time, allowing these values to be treated as constants. While
the array itself is not constant, the values of len and cap can be
determined during the preprocess. This change eliminates the need for
machine.EvalStatic that causes a vm crash.

---------

Co-authored-by: Lee ByeongJun <lbj199874@gmail.com>
Co-authored-by: Petar Dambovaliev <petar.atanasov.1987@gmail.com>