feat: add r/moul/present (#3668)
This realm is a good example of reusing and integrating some of my
recently added libraries:

-  `p/moul/md` #2819 
-  `p/moul/mdtable` #3100 
-  `p/moul/realmpath` #3257 
-  `p/moul/txlink` #3289 
-  `p/moul/collection` #3321 
-  `p/demo/avl/pager` #2584 

It helped me identify that `txlink` was not escaping the arguments,
which resulted in invalid links. (fixed in #3682)

Additionally, it provided me with a better understanding of:

- The shortcomings of the `p/moul/md` API, particularly regarding `"\n"`
handling
-  The need for improved management of the pager for `p/moul/collection`
- What kind of UI improvements we could need on gnoweb. #3355 

Demo: 



https://github.com/user-attachments/assets/4b20cee8-b8d7-4eba-90a8-5b87a3c19521



I also suggest you to look at the `filetest.gno` file.

I believe we should proceed with the merge, to inspire others to create
similar composed realms.

However, I have a few improvement ideas:
1. Extract most of the generic logic into a `p/moul/present`.
2. Consider either making r/moul/present importable by
`r/coreteam/present` to create a hub for presentations from all
teammates, or the opposite: make `r/coreteam/present` the content source
and allow `r/moul/present` to display a subset where `author="moul"`
What are your thoughts?
3. Clean up the code using an improved `p/moul/md` and possibly new `p/`
generic utilities.

Depenes on #3682

---------

Signed-off-by: moul <94029+moul@users.noreply.github.com>