fix(p2p): Avoid inifinty loop during transport/listener `Accept` (#3662)
This PR aims to fix two potential infinite loops during `Accept` in the
`p2p` package.

- First, it removes the `ErrTransportInactive` error. Based on how
`Accept` is used, I believe the inactive error does not make much sense
here. `Accept` should always block, even in an inactive state. Having
`Accept` in two states—blocking and non-blocking—complicates the logic
unnecessarily. Additionally, returning an inactive error creates an
infinite loop, as the handler will keep looping over `Accept` until a
blocking state is reached.

- If the underlying transport or listener is closed, the context should
also be canceled, as this is not a recoverable error. Currently, if the
transport or listener is closed, it will loop indefinitely until the
context is properly closed. While this situation should rarely occur,
since the context should be canceled when closing the transport, I
believe it is safe to force the cancellation of the context to avoid any
potential deadloop.

**EDIT:** I've removed the context select from the loop to simplify the
logic. The logic remains unchanged, but it avoids having two sources of
truth.

This PR also reduces unnecessary error and warning noise when the
listener or transport is closed.

---------

Signed-off-by: gfanton <8671905+gfanton@users.noreply.github.com>