fix: utilize `p2p.ExternalAddress` properly for dialing (#3581)
## Description

This PR fixes a bug with the `Transport` bind address, where the bound
address was taking in a wrong param, as part of an oversight.

It also adds context on the newly introduced checks.

The PR also fixes an issue with the redial mechanism for persistent
peers, where peers wouldn't be redialed more than once.