refactor(rpc): handlers.go - makeJSONRPCHandler() control flow refactor (#3716)
I am proposing a simpler way to write the makeJSONRPCHandler function.
There was a recent bug due to this function that turned lists of
requests with a single item to a object that is a response--but not a
list of responses. With this refactor, this function would be more
resilient to bugs because of its simplified control flow: it first tries
to unmarshal as a slice of requests, and returns an array of responses
if successful. It then tries to unmarshal it into a single request, and
returns a single response if successful. This refactor also abstracts
the logic that processes the requests out to `processRequests(r, request, funcMap, logger)` to see if they actually need to be sent (eg. notifications do not need to be sent). All of that logic is now in the processRequest function, and
improves on the previous implementation because now makeJSONRPCHandle is
much easier to read and understand. If both unmarshals fail, the
function returns an error "error unmarshalling request," same as before.
Please let me know what you think of the refactor. Thanks.