fix(gnolang): make Go2Gno return a prespective error instead of sudden/elusive runtime panic with a bad receiver
The pattern:

```go
func() A()
```
confuses Go into expecting a receiver and it returns a compile time
error "missing receiver", but previously Gno panicked with a runtime
error due to a deference yet the Recv.List was empty. This change
fixes that by detecting that condition and prescriptively panicking
which can then be relayed reasonably as expecting to the calling user.

Fixes #3727