fix(gnolang): prescriptively reject *ast.(IndexListExpr, GoStmt) as unrecognized for Gno
The *ast.IndexListExpr is used for generics but in assignment operations
it is illegal to use. This change returns a proper error and matches
Go's output.
Also *ast.GoStmt is for spawning Go routines but those are forbidden
in Gno, hence reject them prescriptively instead of just spewing
out the raw ast type.

Fixes #3731
Updates #3751