chore(boards2): remove legacy user call asserts (#3756)
Removing these asserts allows realms to interact with the boards which
is convenient when implementing custom permissions/DAO to support custom
features.

An example case could be a realm that publishes tutorials which could
implement the permissions and have a feature where after a tutorial is
published would create a thread within a tutorials board to deal with
discussions or questions related to it.