chore(boards2): simplify default permissions implementation (#3770)
The default permissions implementation is an implementation of the
`Permissions` interfaces that is used by default when creating new
boards and also used as the default realm permissions.

The original idea is to be able to switch the implementation for a
custom one allowing boards to customize the board interaction rules, or
to be able to switch the realm implementation for a different one that
uses another DAO with different board creation rules for example. This
is an idea that still requires further discussions and refinement.

This PR removes the default permission creation options which at this
point makes no sense given that we don't have a boards2 package and
default permissions are only used within the realm.