build: use go1.23.6 (#3767)
Seeing as Go is now at version 1.24, making this PR to bump to the
latest patch version of 1.23. This will also fix the CI on master.

Changes aside from go.mod concern updating a few methods in `txlog`
which were always intended to use 1.23 iterators, but couldn't until
now.

There is a language change (the aforementioned range funcs), but it
shouldn't impact existing code. Pinging those who I think could verify
this in the reviewers.

Fixes #3033.