feat(boards2): allow changing permissions implementation (#3718)
Permissions interface allows assigning different permissions
implementation to the boards2 realms and to individual boards.

This interface may use a DAO to deal with members and governance.

It would be convenient to be able to switch the permissions interface to
allow each board to have a custom implementation that enforces custom
rules.

Apart from individual boards it's also convenient to be able to assign a
custom implementation for boards realm permissions which for example
might want to use a different DAO.