feat(examples): add `p/thox/{accesscontrol,timelock}` (#2307)
We have developed two packages: `accesscontrol` and `timelock` inspired
by openzeppelin contracts. These packages were created in collaboration
with @mous1985 , @DIGIX666 , and myself.

The `accesscontrol` package was primarily designed to support the
development of the `timelock` package, but it can also be used
independently for many other use cases.

### Features

#### Accesscontrol

The `accesscontrol` package provides a library for managing roles and
permissions within Gno. It allows for the creation, assignment, and
management of roles with specific administrative privileges, ensuring
that only authorized accounts can perform certain actions.

#### Timelock

The `timelock` package offers a library for scheduling, canceling, and
executing time-locked operations in Gno. It ensures that operations are
only carried out after a specified delay and provides mechanisms to
manage and verify the status of these operations. The creation of the
`accesscontrol` package was necessary to provide role and permission
management required for the administrative tasks of `timelock`.

### Use Cases

#### Accesscontrol

- Realm Administration Management: Create administrator roles to manage
realms and assign or revoke roles as needed.
- Role-Based Access Control (RBAC): Implement an RBAC system to control
who can access which resources and perform which actions within a Gno
dApp.
- Security and Compliance: Use roles to ensure that only authorized
individuals can perform critical actions, helping to meet security and
compliance regulations.

#### Timelock

- Delayed Transactions: Schedule transactions or actions to be executed
at a specific future time.
- Asset Locking: Implement asset locking mechanisms where users must
wait for a certain period before they can access or move assets.
- Task Automation: Automate periodic or conditional tasks using specific
time delays.

These examples of use cases are not exhaustive, and many other things
are possible with these packages.

<!-- please provide a detailed description of the changes made in this
pull request. -->

<details><summary>Contributors' checklist...</summary>

- [ ] Added new tests, or not needed, or not feasible
- [ ] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [ ] Updated the official documentation or not needed
- [ ] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [ ] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>

---------

Co-authored-by: théo dub <theodub@MacThox.local>
Co-authored-by: mous1985 <mustapha.benazzouz@outlook.fr>
Co-authored-by: Mustapha <102119509+mous1985@users.noreply.github.com>
Co-authored-by: deelawn <dboltz03@gmail.com>
Co-authored-by: Thox <90353329+DIGIX666@users.noreply.github.com>
Co-authored-by: DIGIX666 <digixit66@gmail.com>
Co-authored-by: Manfred Touron <94029+moul@users.noreply.github.com>
Co-authored-by: Morgan <git@howl.moe>
Co-authored-by: Morgan <morgan@morganbaz.com>