feat(gnovm/tests/stdlibs): add `fmt` (#3847)
This PR removes `fmt` from being implemented using NativeValue /
NativeType, and implements a mostly pure-Gno implementation which uses
an initial, rudimentary implementation of Gno reflection for `fmt`.

This PR is ready for review, but some of its related work has been
broken out in separate PR to aid reviewing and create atomic PRs.

- Blocks #3677
- Relates to #1361
- Depends on #3861
- Depends on #3862 
- Depends on #3875

## Summary of changes

### addition of `gnovm/tests/stdlibs/fmt`

The main addition is of this set of files, which comprises the bulk of
the added lines of this PR. Most of this is a copy of Go's `fmt`.
Changes are mostly in `errors.gno` and `format.gno` - the other files
are left mostly intact.

At the end of `print.gno`, there is an added section containing a set of
native functions which form the "reflection" needed to implement the
`fmt` package in a manner that is similar to Go's. The implementation is
not meant to be a public API and is meant to serve initially only for
usage by `fmt`; we can use this as sketch to then implement a proper
`reflect` package.

Attention in reviewing should mostly go to this `print.gno` and its
`print.go` counterpart, especially for any missed cases or potential
bugs.

### addition of `gnovm/tests/stdlibs/os`

Complementary to the above, to avoid some problems that were arising due
to the handling of os.Stdin, Stdout and Stderr, I also removed the
NativeValue / NativeType in `os` and replaced them with a native binding
alternative.

To avoid having both an `Output` and `ErrorOutput` on the machine, this
package supports an additional method on the `Machine.Output`,
`StderrWrite`, which is implemented by writers who wish to differentiate
between the two.

### other

- `gnovm/pkg/test`
	- added support for `StderrWrite` in its writers
	- removed native values for `os`  and `fmt`
- combine `stdin`, `stdout`, `stderr` into a single `output` parameter
in the related functions.
- `gnovm/pkg/gnolang`
- support testing `gnovm/tests/stdlibs`, by only testing stdlibs that
exist there and not in the official stdlibs.
- `gnovm/cmd/gno`
	- fix tests which needed the `fmt` package.
- `gnovm/tests/files`
- most changed tests are result of the improved functionality of the
`fmt` package, allowing for instance to show declared types correctly.
- `gnovm/stdlibs/bufio`
	- due to the removal of `os.Stdin`, this example had to be changed.
- `misc/genstd`
- added a flag `-skip-init-order` when creating it could cause problems,
like the testing stdlibs; fixed a bug when using `TypedValue`.