feat: lock token transfer and parameter module, take 2 (#3860)
This is a continuation of https://github.com/gnolang/gno/pull/3176.

It breaks the params struct into fields and stores them in
<module>:_:<fieldname>. All parameters are of the form <module>:<name>
or <module>:<realm/_/*>:<name>. A colon is used rather than a period/dot
because it's ambiguous with realm domains.

std.SetParam is realm-local only, limited and safe.
sys/params.SetSysParam is arbitrary and global. Need to make sure it's
only importable by r/sys/params, but I think we said we'd do this with
"internal" or something. Maybe for a followup PR.

The ParamsKeeper is divided into prefix spaces, `pk.WithModule("bank")`
for example, and that is passed into the BankKeeper, that way the
BankKeeper only has access to the bank:* param space. "language
capabilities" security. But the VMKeeper needs to the root ParamsKeeper
because it needs to be able to modify anything via r/sys/params.

I got rid of the .type suffix of param keys. It's less safe right now
but we can add type safety to this later. You *can* and *must* use the
type suffix from toml for certain types though; "strings" which
represent []strings must be suffixed by .strings, because toml only
gives []interface{}{}.

---------

Co-authored-by: piux2 <90544084+piux2@users.noreply.github.com>
Co-authored-by: Manfred Touron <94029+moul@users.noreply.github.com>
Co-authored-by: Morgan <morgan@morganbaz.com>