feat: add assets.gnoteam.com to the CSP img-src  whitelist (#3892)
In order to allow gnoteam images to be served through CSP, this PR adds
the dedicated domain https://assets.gnoteam.com/ to the rules. In
addition to https://gnolang.github.io/, which is used for the blog
system, gnoteam assets will also be used for general purposes, ensuring
secure and consistent asset delivery.