fix(examples): a p/moul/authz.ContractAuthority could execute multiple actions (#3913)
There was a once.Once object that prevented one action to be executed
twice by the ContractAuthority handler. But it also prevented the
ContractAuthority from executing any other actions.

The fix was to link the once.Once object to the action, not to the
ContractAuthority.
A test has also been added to test the case.

---------

Signed-off-by: D4ryl00 <d4ryl00@gmail.com>
Co-authored-by: Manfred Touron <94029+moul@users.noreply.github.com>