fix(pkg/doc): include Consts and Vars of the type in JSON (#3912)
This is a follow-on to the vm/qdoc PR
https://github.com/gnolang/gno/pull/3459 . We already handle the case
where the AST lists type constructors as part of the Type description.
We also need to handle the case where consts and vars of the type are
listed as part of the Type description (not in the main Consts and Vars
sections). See the [updated lines in the hello
test](https://github.com/gnolang/gno/blob/a693cd5146591261f90956e1899f5460d39535bc/gnovm/pkg/doc/testdata/integ/hello/hello.gno#L14-L18).
```
  var myStructPtr *myStruct
  myStructConst *myStruct = &myStruct{a: 1000}
```
For these values to appear in the vm/qdoc result, this PR fix is needed.

We also fix an oversight in the `JSONField` struct to add the json field
names.

---------

Signed-off-by: Jeff Thompson <jeff@thefirst.org>