feat: simplify tally allowing prop definitions to deal with logic
Previous implementation of tally counted the votes and checked the that
a quorum was reached. That changed to allow devs to implement all the
tally logic within the proposal definition. This simplifies the
`ProposalDefinition` interface and the CommonDAO type while at the same
time allows for more flexibility and a less opinionated tally
implementation.