fix(gnovm): charge gas for type checking the Go AST statically
Typechecking the Go abstract syntax tree incurs a huge penalty
as part of the GnoVM's functionality when starting up and before
running a package, hence it is only natural to also charge a
cost. At first the charges shall be naive and as we expend more
thought and research, shall refine what and how much to charge.

Fixes #3560