feat(boards2): add realm and realm members locking support  (#3969)
PR adds support to `boards2` realm so it can be locked.

Realm can be locked leaving the possibility to manage realm members
(council DAO) open or otherwise can be fully locked disallowing any
member changes.

Support for setting a global notice message is also part of this PR.
Notice can be used as a standalone feature or in conjunction with the
locking functionality to inform users of the location of the new realm
version.

Part of #3494

Also see
[comment](https://github.com/gnolang/gno/issues/3494#issuecomment-2719739571)