feat(examples/hof): Title and description for realms, sorting by upvotes and newest (#3674)
# Title And Description for Hall of Fame realms

## Goal
Add title and description to realms in the Hall of Fame, as well as
functionality of sorting realms by upvotes and by date of publishing.

## Improvements

### Title and Description
Realms in Hall of Fame will now have title and description, this will
give users more context about each realm.
From now on, people who wish to publish their realm in the Hall of Fame
will have to provide a title and description when registering their
realm.

### Sorting Realms by Upvotes and Newest realms
Realms in the Hall of Fame will now have the functionality to be sorted
by upvotes and most recent realms. This will improve user experience by
making it easier to find specific realms or explore new ones.

## **Feature Progress** 
- [x] Title and description functionality  
- [x] Tests for title and description functionality  
- [x] Sorting realms by upvotes and newest realms functionality  
- [x] Tests for sorting realms by upvotes and newest realms
functionality

---------

Co-authored-by: Ivan Ursulovic <ivan@pop-os.localdomain>
Co-authored-by: Ivan Ursulovic <ivan@Ivans-MacBook-Air.local>
Co-authored-by: Leon Hudak <33522493+leohhhn@users.noreply.github.com>