chore: revert "feat(gnoweb): add metadata to md content (#3797)" (#3924)
This reverts commit d1db75eb9250645b61344cdaa651f361e30bd60d (#3797).
See #3910.

Let's revert this now. Then we can try to improve SEO automatically by
extracting the first H1 and the first paragraph, as this PR did
manually. After that, we should take more time to consider aspects such
as frontmatter and everything involved in allowing developers to specify
"invisible" elements.