ci: In dependabot.yml, use the label "dependencies" everywhere (#3982)
As we discussed in the review team meeting, dependabot [gives the
warning](https://github.com/gnolang/gno/pull/3954#issuecomment-2727909534)
"The following labels could not be found: `github_actions`." The weekly
schedule uses the label `dependencies`. This PR updates the daily
schedule to also use this label. This should remove the warning.

Signed-off-by: Jeff Thompson <jeff@thefirst.org>