fix: gnogenesis generate subcommand (#4021)
The `gnogenesis generate` subcommand no longer allows generating a valid
genesis. Starting a gnoland node using it produces the following panic:
`panic: auth genesis state cannot be empty`

The issue is that changes introduced
[here](https://github.com/gnolang/gno/blob/d24efa034173cb57168e27afdb3dcce4a07767fb/tm2/pkg/sdk/auth/genesis.go#L25-L32)
now require the `auth` section of the `app_state` to not be empty. The
following fix simply sets the default values for the `app_state` of the
genesis generated by `gnogenesis generate`.