feat(gnovm): custom go parser for gas counting and nesting limitation
The Go parser from the standard library has been imported in
gnovm/pkg/parser, with a Makefile to generate and apply custom
patch.

The patch consist to add a new function ParseFileStats which performs
accounting to enable gas computation and set the maximum nesting level
as an input parameter to instead of a constant.

Parsing is instrumented to measure and control the nesting level and
have a first cost estimate of parsing to compute gas, before type
checking itself.

The advantages are that there is almost no overhead, early catch of
problems (even before the AST memory is allocated), and a simple gas
formula (no distinction between tokens except the nesting level). No
extra pass is added, we just exploit some internal parser metadata for
prevention of too deep nesting and gas counting.

The drawback is to fork and maintain a different version of the Go
parser, but most of it is automated.

Note: the Go parser has been imported from go1.24.1, as it has been
simplified and has no internal dependencies, as opposite to previous
versions.