feat: token hub (#3479)
### Description

Token hub is a registry for tokens and NFTs created on gno.land. Since
all the tokens are visible in each their own realms, as it is right now
there is no easy way that a user can check which tokens belong to him by
his address. Token hub bypasses this by concentrating each each token in
it's own realm and allows operations over these tokens providing a
easier way to communicate with the realms in which the tokens are
originally made. Since grc20 registry already exists the idea is to
combine it's functionality with similar concept used for other tokens to
create a realm that will contain information about all tokens, this way
you will be able to get all your token balances in one place.

<details>
  <summary>See more details</summary>

### How it works
In the realm where token is being made, tokenhub needs to be imported
and the newly made token has to be registered to the token hub by using
RegisterToken function and passing the token as an argument. This is
everything token hub needs to be able to operate. Simiraly how it is
done
[here](https://github.com/gnolang/gno/blob/master/examples/gno.land/r/demo/grc20reg/grc20reg.gno)
for GRC20 token hub registers a pointer to the token in it's registry.
- For GRC20 the mentioned registry is used in Token Hub
- For GRC721 similar approach as in that registry is implemented; note:
each NFT has to be registered after minting because each NFT has his own
ID. Although nft collection could be registered without keeping track of
IDs, this would prove to be useless for anything besides statistics (how
many NFTs are minted in that collection)
- For GRC1155 the method is a bit different - since multiple fungible
and non-fungible tokens are a part of the same collection, to keep track
of balances one would need both collection (getter) and the token ID,
this is why in this case tokens of the same collection are registered
separately, which at the end of the day makes sence, but does represent
an extra step both in registration and in getting the balance

### More Details
In order to make this work I had to make minor additions to other realms
as one would probably notice in the code - I had to add Getters
(pointers) in the grc721 and grc1155 so storing the token instance
itself is avoided, like it is done with GRC20 for grc20reg.gno purposes.
I also needed to add Iterate and IterateN functions to grc20 only to be
able to get the tokens of that registry which shouldn't undermine any of
it's functionality or security.
</details>

<details>
  <summary>See screenshots</summary>


![image](https://github.com/user-attachments/assets/cc18b1b3-ec90-45f5-9ed5-a29b80678c3f)

![image](https://github.com/user-attachments/assets/2098f4dc-edc1-4edf-8135-7c941f058677)

![image](https://github.com/user-attachments/assets/7420cdf3-7e53-46ec-905a-ec0aa4e96c4c)
</details>

All feedback is welcome

---------

Co-authored-by: Leon Hudak <33522493+leohhhn@users.noreply.github.com>