chore: Add misc/stress-test (#1583)
This PR adds a Go utility in misc/stress-test/stress-test-many-posts
which adds millions of posts to the (local) boards realm to test how
memory and transaction time change with lots of realm storage. See the
README for how to use it. See [this
report](https://github.com/gnolang/hackerspace/issues/67) for example
graphs created from using this utility.

---------

Signed-off-by: Jeff Thompson <jeff@thefirst.org>