feat: gas fee estimation (#4102)
## ISSUES

We observed very high gas fee usage in the test network. Here is what
happened.

 The network charges users the gas fee they specify when they submit a
transaction—regardless of whether the transaction eventually fails or
succeeds.A transaction may fail for several reasons, including
insufficient gas, an incorrect function name, an improper parameter
type, or invalid values that cause the transaction to fail. As long as
the gasFee/gasWanted exceeds the gasPrice, along with passing other
basic validations, the transaction will be allowed into the mempool.


At that point, fees are deducted from the user’s account even though the
transaction has not yet been executed on the virtual machine. If the
transaction later fails during the execution, the chain still retains
the fee. This design allows for basic validation upfront and defers the
full execution check until block playback.

However, a problem arises because users must estimate both the gas to be
used and the required fee. They tend to overestimate the gas limit and
gas fee to ensure that the transaction goes through,  and when a
transaction fails, the actual gas consumed is typically much lower than
the estimation.

As a result, when the third party wallet calculates GasFee/GasUsed, the
resulting figure can be significantly larger than the gas price set
on-chain in each block. Then the wallet suggests the gas fee by applying
the formula GasFee/GasUsed multiplied by simulated gas, the resulting
gas fee can be even higher. This is why we sometimes see an
overestimated value. The solution below should provide a more accurate
estimation of the fee required on the network

## SOLUTION

1. We provide an on-chain query to retrieve the latest gas price,
instead of calculating it using the GasFee / GasUsed ratio. Third-party
wallets can use the on-chain gas price multiplied by the simulated gas
to estimate the GasFee for users.

$ gnokey query auth/gasprice
height: 0
data: {
  "gas": "1000",
  "price": "1ugnot"
}

2. We provide an estimated gas fee when users use gnokey with the
‘-simulation only’ flag. There are no charges incurred during the
simulation.
```

   gnokey maketx call -pkgpath gno.land/r/hello -func Hello -gas-wanted 2000000 -gas-fee 1000000ugnot -broadcast -chainid tendermint_test -simulate only test1
        [stdout]
        ("hello" string)
        OK!
        GAS WANTED: 2000000
        GAS USED:   98577
        HEIGHT:     0
        EVENTS:     []
        INFO:       estimated gas 98577, gas fee: 103ugnot, current gas price: 1000gas/1ugnot
```

Then users can use estimated gas fee and gas used number in the
simulation to submit a transaction.
```

   gnokey maketx call -pkgpath gno.land/r/hello -func Hello -gas-wanted 98577 -gas-fee 103ugnot -broadcast -chainid tendermint_test test1
        [stdout]
        ("hello" string)


        OK!
        GAS WANTED: 98577
        GAS USED:   98527
        HEIGHT:     53
        EVENTS:     []
        INFO:
        TX HASH:    20+++XrLH81fR53iknyujpQXJWQstDrT13nZ6iTRk/o=


```
3. We added telemetry metrics for any OTEL compatible tools to monitor
gas price in real time.

![Pasted Graphic
54](https://github.com/user-attachments/assets/d68b25a6-bf08-4375-88f3-034c2e704367)
![Pasted Graphic
53](https://github.com/user-attachments/assets/6eaf0a23-d920-4970-93f1-3d828103bc8c)

---------

Co-authored-by: Manfred Touron <94029+moul@users.noreply.github.com>