fix: github-bot multiple fixes on ReviewBy requirements (#4145)
This PR:
-
https://github.com/gnolang/gno/commit/f99df53babbb496691df264b05d2fadb7dcb673b
Improves the `gnoweb` review rule to make it fit with [this
description](https://github.com/gnolang/gno/issues/3238#issuecomment-2793635486):
> a. If Alexis or Guilhem is the author of a PR related to gnoweb, the
other is assigned as a reviewer.
b. If someone else is the author of a PR related to gnoweb, both are
assigned as reviewers, but only one review from one of them is required.
-
https://github.com/gnolang/gno/commit/7088bffb4e5a06c4e0b6ccf8350cd90b8f70edc3
Prevents [an
error](https://github.com/gnolang/gno/issues/3238#issuecomment-2783717659)
when the bot try to request a review from the PR author.
-
https://github.com/gnolang/gno/commit/24b473848776f9c63322718868ef67e43e3b3e3e
Adds a `RequestAction` mechanism similar to the `LabelAction` to the
`ReviewByUser` and `ReviewByTeam` requirements.